Adds some suggestions
There were a few small issues that were causing the problems you encountered, but they mostly came from the fact that you had so much copy-paste code. I tried to consolidate that in to a few functions, and it seems to be working much better now. Let me know if I can help with anything else.